Update @yoast/components prepublishOnly command
This ensures that the published package has the same structure as a locally linked monorepo. This makes it easier to import non-JS dependencies from the monorepo since they don't allow to import 'from the root'. This means that we have to use a file import, which fails on either production or development because the published package does not match the development structure