Improve code style and remove markings from research
I changed the structure of the research research result. Now, the whole research uses the same object structure where each keyword is represented in one object with its forms and occurrences. This object structure is also returned as the result. Additionally, markings have been removed from the research and added to the assessment.